Swap the contents of `DatasetTests` and `DataSetTests_Integration` and also fix capitalization.
